feat: useInventoryHistory hook
